check_azure_rbac_permission_mismatch/requirements.txt
azure-identity>=1.10.0
requests>=2.25.0